+++ projects/ngx-contextmenu/src/lib/services/context-menu/context-menu.service.ts
@@ -137,7 +137,9 @@ export class ContextMenuService {
       const positionStrategy = this.overlay
         .position()
         .flexibleConnectedTo(
-          new ElementRef(event ? event.target : anchorElement)
+          new ElementRef(
+            event instanceof MouseEvent ? event.target : anchorElement
+          )
         )
         .withPositions([
           {